chapter8/requirements.txt
protobuf==5.26.0
